+++ src/prophetverse/engine/vi.py
@@ -3,11 +3,12 @@
 The VIInferenceEngine class performs Variational Inference using SVI with
 configurable autoguides specified as string parameters.
 """
-
-from typing import Optional
+import warnings
+from typing import Optional, Literal
 
 import jax.numpy as jnp
 import numpyro
+from numpyro.handlers import condition
 from numpyro.infer import SVI, Trace_ELBO
 from numpyro.infer.autoguide import (
     AutoDiagonalNormal,
@@ -17,6 +18,7 @@ from numpyro.infer.autoguide import (
 )
 from numpyro.infer.initialization import init_to_mean
 from numpyro.infer.svi import SVIRunResult
+from jax.random import split
 
 from prophetverse.engine.base import BaseInferenceEngine
 from prophetverse.engine.optimizer.optimizer import (
@@ -34,6 +36,34 @@ GUIDE_MAP = {
     "AutoDiagonalNormal": AutoDiagonalNormal,
     "AutoLowRankMultivariateNormal": AutoLowRankMultivariateNormal,
 }
+Guides = Literal["AutoNormal", "AutoMultivariateNormal", "AutoDiagonalNormal", "AutoLowRankMultivariateNormal", "AutoDelta"]
+
+
+def _fit_svi(
+    rng_key,
+    model,
+    guide,
+    optimizer,
+    num_steps,
+    progress_bar,
+    stable_update,
+    forward_mode_differentiation,
+    **kwargs,
+) -> SVIRunResult:
+    svi_ = SVI(
+        model,
+        guide,
+        optimizer,
+        loss=Trace_ELBO(),
+    )
+    return svi_.run(
+        rng_key=rng_key,
+        progress_bar=progress_bar,
+        stable_update=stable_update,
+        num_steps=num_steps,
+        forward_mode_differentiation=forward_mode_differentiation,
+        **kwargs,
+    )
 
 
 class VIInferenceEngine(BaseInferenceEngine):
@@ -77,7 +107,7 @@ class VIInferenceEngine(BaseInferenceEngine):
 
     def __init__(
         self,
-        guide: str = "AutoNormal",
+        guide: Guides = "AutoNormal",
         optimizer: Optional[BaseOptimizer] = None,
         num_steps=10_000,
         num_samples=_DEFAULT_PREDICT_NUM_SAMPLES,
@@ -130,33 +160,7 @@ class VIInferenceEngine(BaseInferenceEngine):
         """
         self.guide_ = self._guide_class(self.model_, init_loc_fn=self._init_loc_fn)
 
-        def get_result(
-            rng_key,
-            model,
-            guide,
-            optimizer,
-            num_steps,
-            progress_bar,
-            stable_update,
-            forward_mode_differentiation,
-            **kwargs,
-        ) -> SVIRunResult:
-            svi_ = SVI(
-                model,
-                guide,
-                optimizer,
-                loss=Trace_ELBO(),
-            )
-            return svi_.run(
-                rng_key=rng_key,
-                progress_bar=progress_bar,
-                stable_update=stable_update,
-                num_steps=num_steps,
-                forward_mode_differentiation=forward_mode_differentiation,
-                **kwargs,
-            )
-
-        self.run_results_: SVIRunResult = get_result(
+        self.run_results_: SVIRunResult = _fit_svi(
             self.rng_key,
             self.model_,
             self.guide_,
@@ -191,7 +195,7 @@ class VIInferenceEngine(BaseInferenceEngine):
         """
         losses = run_results.losses
         if jnp.isnan(losses)[-1]:
-            msg = "NaN losses in VIInferenceEngine."
+            msg = f"NaN losses in '{self.__class__.__name__}'."
             msg += " Try decreasing the learning rate or changing the model specs."
             msg += " If the problem persists, please open an issue at"
             msg += " https://github.com/felipeangelimvieira/prophetverse"
@@ -219,6 +223,48 @@ class VIInferenceEngine(BaseInferenceEngine):
         )
         self.samples_ = predictive(rng_key=self.rng_key, **kwargs)
         return self.samples_
+
+    def _update(self, site_names, mode="mean", **kwargs):
+        assert (
+            self.posterior_samples_ is not None
+        ), "Can only update from a fitted instance!"
+
+        if mode != "mean":
+            warnings.warn(
+                "Only 'mean' mode is supported for MAPInferenceEngine. Ignoring the provided mode."
+            )
+
+        to_condition_on = {
+            k: v for k, v in self.posterior_samples_.items() if k not in site_names
+        }
+
+        conditioned_model = condition(self.model_, data=to_condition_on)
+        temp_guide = self._generate_guide(conditioned_model)
+        rng_key, _ = split(self.rng_key)
+
+        # NB: not sure if we should overwrite, keep or simply discard the results?
+        temp_results: SVIRunResult = _fit_svi(
+            rng_key,
+            conditioned_model,
+            temp_guide,
+            self._optimizer.create_optimizer(),
+            self._num_steps,
+            stable_update=self.stable_update,
+            progress_bar=self.progress_bar,
+            forward_mode_differentiation=self.forward_mode_differentiation,
+            **kwargs,
+        )
+
+        self.raise_error_if_nan_loss(temp_results)
+
+        updated_posterior = temp_guide.sample_posterior(
+            self.rng_key, params=temp_results.params, **kwargs
+        )
+
+        updated_posterior.update(to_condition_on)
+        self.posterior_samples_ = updated_posterior
+
+        return self
 
     @classmethod
     def get_test_params(*args, **kwargs):
